refactor: 시스템 로그 최종 수정 및 Sidebar 엔드포인트 수정
